requirements.txt
numpy==1.21.2
opencv-python==4.5.3.56
keyboard==0.13.5
pillow==8.3.1
pygame==2.0.1
pyautogui==0.9.53
pyscreenshot==2.2
pynput==1.7.1
pytesseract==0.3.8
scikit-learn==0.24.2
pandas==1.3.3